requirements.txt
-e git+https://github.com/SohierDane/BigQuery_Helper@8615a7f6c1663e7f2d48aa2b32c2dbcb600a440f#egg=bq_helper
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
google-api-core==1.14.3
google-auth==1.8.2
google-cloud-bigquery==1.23.0
google-cloud-core==1.1.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
idna==2.8
numpy==1.17.4
pandas==0.25.3
plotly==4.4.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
retrying==1.3.3
rsa==4.0
six==1.13.0
urllib3==1.25.7
